Use BlockBindingsSourceMenu instead of BlockBindingsMenuContent
